# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alex production",
        "description": "Documentation for alex",
        "contact": {
            "email": "alex@gmail"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/author": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get list of Author",
                "description": "Returns list of Author",
                "operationId": "indexAuthor",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Insert a new Author",
                "description": "Insert a new Author",
                "operationId": "storeAuthor",
                "requestBody": {
                    "description": "Author to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/author/{author_id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Show a Author from his Id",
                "description": "Show a Author from his Id",
                "operationId": "showAuthor",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Author--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete a Author",
                "description": "Delete a Author",
                "operationId": "destroyAuthor",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Author--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update an existing Author",
                "description": "Update an existing Author",
                "operationId": "updateAuthor",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Author--id"
                    }
                ],
                "requestBody": {
                    "description": "Author to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Author": {
                "title": "Author",
                "description": "Author model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Author",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Author",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Author",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Author",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "Author--id": {
                "name": "Author_id",
                "in": "path",
                "description": "Id of Author",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        }
    }
}